Move the react-native yaob setup to its own file
